pyproject.toml
[project]
name = "ingestor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiokafka>=0.13.0",
    "fastapi[standard]>=0.135.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pytest-cov>=7.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.4.0",
]
[project.optional-dependencies]
dev = [
    "editorconfig-checker==3.2.1",
    "pre-commit>=4.3.0",
    "ruff==0.11.13",
]
